.Net Core 1.1 : already an open DataReader
"InvalidOperationException: There is already an open DataReader associated with this Command which must be closed first."

It seems like if we send an IQueryable to the View, there will be then a problem with the taghelper "RoleUserTagHelper.cs".
IdentityRole role = await roleManager.FindByIdAsync(Role); => will throw the error.

This solution works. But i don't know if it's a normal behavior or a larger bug of .Net Core and identity core version.